+++ tests/dial/augend/hexcolor_spec.lua
@@ -3,9 +3,9 @@ local hexcolor = require("dial.augend").hexcolor
 describe("Test of hex colors", function()
     describe("config", function()
         describe("case", function()
-            it('"keep" is the default', function()
+            it('"prefer_lower" is the default', function()
                 local augend = hexcolor.new()
-                assert.are.same(augend.config, { case = "keep" })
+                assert.are.same(augend.config, { case = "prefer_lower" })
             end)
 
             it('"upper" is accepted', function()
@@ -16,6 +16,16 @@ describe("Test of hex colors", function()
             it('"lower" is accepted', function()
                 local augend = hexcolor.new { case = "lower" }
                 assert.are.same(augend.config, { case = "lower" })
+            end)
+
+            it('"prefer_upper" is accepted', function()
+                local augend = hexcolor.new { case = "prefer_upper" }
+                assert.are.same(augend.config, { case = "prefer_upper" })
+            end)
+
+            it('"prefer_lower" is accepted', function()
+                local augend = hexcolor.new { case = "prefer_lower" }
+                assert.are.same(augend.config, { case = "prefer_lower" })
             end)
 
             it("rejects other values", function()
@@ -87,50 +97,51 @@ describe("Test of hex colors", function()
     end)
 
     describe("add", function()
-        describe("case = keep", function()
-            local augend = hexcolor.new { case = "keep" }
+        describe("case = upper", function()
+            local augend = hexcolor.new { case = "upper" }
 
             it("can increment the red hex value", function()
                 augend.kind = "r"
                 assert.are.same(augend:add("#000000", 001), { text = "#010000", cursor = 3 })
                 assert.are.same(augend:add("#000000", 016), { text = "#100000", cursor = 3 })
-                assert.are.same(augend:add("#000000", 255), { text = "#ff0000", cursor = 3 })
-                assert.are.same(augend:add("#000000", 256), { text = "#ff0000", cursor = 3 })
+                assert.are.same(augend:add("#000000", 255), { text = "#FF0000", cursor = 3 })
+                assert.are.same(augend:add("#000000", 256), { text = "#FF0000", cursor = 3 })
             end)
 
             it("can increment the green hex value", function()
                 augend.kind = "g"
                 assert.are.same(augend:add("#000000", 001), { text = "#000100", cursor = 5 })
                 assert.are.same(augend:add("#000000", 016), { text = "#001000", cursor = 5 })
-                assert.are.same(augend:add("#000000", 255), { text = "#00ff00", cursor = 5 })
-                assert.are.same(augend:add("#000000", 256), { text = "#00ff00", cursor = 5 })
+                assert.are.same(augend:add("#000000", 255), { text = "#00FF00", cursor = 5 })
+                assert.are.same(augend:add("#000000", 256), { text = "#00FF00", cursor = 5 })
             end)
 
             it("can increment the blue hex value", function()
                 augend.kind = "b"
                 assert.are.same(augend:add("#000000", 001), { text = "#000001", cursor = 7 })
                 assert.are.same(augend:add("#000000", 016), { text = "#000010", cursor = 7 })
-                assert.are.same(augend:add("#000000", 255), { text = "#0000ff", cursor = 7 })
-                assert.are.same(augend:add("#000000", 256), { text = "#0000ff", cursor = 7 })
+                assert.are.same(augend:add("#000000", 255), { text = "#0000FF", cursor = 7 })
+                assert.are.same(augend:add("#000000", 256), { text = "#0000FF", cursor = 7 })
             end)
 
             it("can increments all hex values", function()
                 augend.kind = "all"
                 assert.are.same(augend:add("#000000", 001), { text = "#010101", cursor = 1 })
                 assert.are.same(augend:add("#000000", 016), { text = "#101010", cursor = 1 })
-                assert.are.same(augend:add("#000000", 255), { text = "#ffffff", cursor = 1 })
-                assert.are.same(augend:add("#000000", 256), { text = "#ffffff", cursor = 1 })
+                assert.are.same(augend:add("#000000", 255), { text = "#FFFFFF", cursor = 1 })
+                assert.are.same(augend:add("#000000", 256), { text = "#FFFFFF", cursor = 1 })
             end)
 
-            it("keeps the casing of the hex values", function()
+            it("converts to upper case", function()
                 augend.kind = "all"
-                assert.are.same(augend:add("#0a1bff", 1), { text = "#0b1cff", cursor = 1 })
+                assert.are.same(augend:add("#0a1bff", 1), { text = "#0B1CFF", cursor = 1 })
+                assert.are.same(augend:add("#0a1BFf", 1), { text = "#0B1CFF", cursor = 1 })
                 assert.are.same(augend:add("#0A1BFF", 1), { text = "#0B1CFF", cursor = 1 })
             end)
 
-            it("uses lower case when initial value has no letter", function()
+            it("uses upper case when initial value has no letter", function()
                 augend.kind = "all"
-                assert.are.same(augend:add("#059799", 5), { text = "#0a9c9e", cursor = 1 })
+                assert.are.same(augend:add("#059799", 5), { text = "#0A9C9E", cursor = 1 })
             end)
         end)
 
@@ -172,6 +183,7 @@ describe("Test of hex colors", function()
             it("converts to lower case", function()
                 augend.kind = "all"
                 assert.are.same(augend:add("#0a1bff", 1), { text = "#0b1cff", cursor = 1 })
+                assert.are.same(augend:add("#0a1BFf", 1), { text = "#0b1cff", cursor = 1 })
                 assert.are.same(augend:add("#0A1BFF", 1), { text = "#0b1cff", cursor = 1 })
             end)
 
@@ -181,8 +193,8 @@ describe("Test of hex colors", function()
             end)
         end)
 
-        describe("case = upper", function()
-            local augend = hexcolor.new { case = "upper" }
+        describe("case = prefer_upper", function()
+            local augend = hexcolor.new { case = "prefer_upper" }
 
             it("can increment the red hex value", function()
                 augend.kind = "r"
@@ -216,15 +228,82 @@ describe("Test of hex colors", function()
                 assert.are.same(augend:add("#000000", 256), { text = "#FFFFFF", cursor = 1 })
             end)
 
-            it("converts to upper case", function()
+            it("keeps lower case if existing color has lower case letters", function()
                 augend.kind = "all"
-                assert.are.same(augend:add("#0a1bff", 1), { text = "#0B1CFF", cursor = 1 })
+                assert.are.same(augend:add("#0a1bff", 1), { text = "#0b1cff", cursor = 1 })
+            end)
+
+            it("uses upper case when initial value has upper case letters", function()
+                augend.kind = "all"
                 assert.are.same(augend:add("#0A1BFF", 1), { text = "#0B1CFF", cursor = 1 })
             end)
 
             it("uses upper case when initial value has no letter", function()
                 augend.kind = "all"
+                assert.are.same(augend:add("#059799", 1), { text = "#06989A", cursor = 1 })
                 assert.are.same(augend:add("#059799", 5), { text = "#0A9C9E", cursor = 1 })
+            end)
+
+            it("converts to upper case if existing color has mixed casing", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#0a1BFf", 1), { text = "#0B1CFF", cursor = 1 })
+            end)
+        end)
+
+        describe("case = prefer_lower", function()
+            local augend = hexcolor.new { case = "prefer_lower" }
+
+            it("can increment the red hex value", function()
+                augend.kind = "r"
+                assert.are.same(augend:add("#000000", 001), { text = "#010000", cursor = 3 })
+                assert.are.same(augend:add("#000000", 016), { text = "#100000", cursor = 3 })
+                assert.are.same(augend:add("#000000", 255), { text = "#ff0000", cursor = 3 })
+                assert.are.same(augend:add("#000000", 256), { text = "#ff0000", cursor = 3 })
+            end)
+
+            it("can increment the green hex value", function()
+                augend.kind = "g"
+                assert.are.same(augend:add("#000000", 001), { text = "#000100", cursor = 5 })
+                assert.are.same(augend:add("#000000", 016), { text = "#001000", cursor = 5 })
+                assert.are.same(augend:add("#000000", 255), { text = "#00ff00", cursor = 5 })
+                assert.are.same(augend:add("#000000", 256), { text = "#00ff00", cursor = 5 })
+            end)
+
+            it("can increment the blue hex value", function()
+                augend.kind = "b"
+                assert.are.same(augend:add("#000000", 001), { text = "#000001", cursor = 7 })
+                assert.are.same(augend:add("#000000", 016), { text = "#000010", cursor = 7 })
+                assert.are.same(augend:add("#000000", 255), { text = "#0000ff", cursor = 7 })
+                assert.are.same(augend:add("#000000", 256), { text = "#0000ff", cursor = 7 })
+            end)
+
+            it("can increments all hex values", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#000000", 001), { text = "#010101", cursor = 1 })
+                assert.are.same(augend:add("#000000", 016), { text = "#101010", cursor = 1 })
+                assert.are.same(augend:add("#000000", 255), { text = "#ffffff", cursor = 1 })
+                assert.are.same(augend:add("#000000", 256), { text = "#ffffff", cursor = 1 })
+            end)
+
+            it("keeps upper case if existing color has upper case letters", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#0A1BFF", 1), { text = "#0B1CFF", cursor = 1 })
+            end)
+
+            it("uses lower case when initial value has lower case letters", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#0a1bff", 1), { text = "#0b1cff", cursor = 1 })
+            end)
+
+            it("uses lower case when initial value has no letter", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#059799", 1), { text = "#06989a", cursor = 1 })
+                assert.are.same(augend:add("#059799", 5), { text = "#0a9c9e", cursor = 1 })
+            end)
+
+            it("converts to lower case if existing color has mixed casing", function()
+                augend.kind = "all"
+                assert.are.same(augend:add("#0a1BFf", 1), { text = "#0b1cff", cursor = 1 })
             end)
         end)
     end)